applications/storium-modulith/requirements.txt
# Django Core
Django>=5.0,<6.0

# DTO / validation
pydantic>=2,<3
Pillow>=10.0          # ImageField desteği için

# Veritabanı (Production)
psycopg2-binary>=2.9  # PostgreSQL sürücüsü

# Environment
python-decouple>=3.8  # .env okuma (veya django-environ)

# Statik Dosya (Production)
whitenoise>=6.6       # Statik dosyaları serve etmek için

# WSGI (Docker / production)
gunicorn>=22.0

# Önbellek (modulith: Redis)
django-redis>=5.4
redis>=5.0

# Geliştirme Araçları (Opsiyonel)
django-debug-toolbar>=4.0
